+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Allow soundboard during postgame
   - Notes modal: fix max size
   - Forbid language change for players
+  - Clear notes when starting a new game
 - Localization:
 
 - Maintenance: